+++ samples/test/sdl32/src/Game.cpp
@@ -305,7 +305,7 @@ void Game::createDescriptorSets() {
 // Loader Models
 //---------------------------------------------
 
-void Game::updateModel(int modelId, glm::mat4 newModel) {
+void Game::updateModel(size_t modelId, glm::mat4 newModel) {
 
     if (modelId >= this->modelList.size()) {
         return;
@@ -314,7 +314,7 @@ void Game::updateModel(int modelId, glm::mat4 newModel) {
     this->modelList[modelId].setModel(newModel);
 }
 
-int Game::createMeshModel(const std::string& modelFile) {
+size_t Game::createMeshModel(const std::string& modelFile) {
     // Import model "scene"
     Assimp::Importer importer;
 
@@ -343,7 +343,7 @@ int Game::createMeshModel(const std::string& modelFile) {
             std::shared_ptr<ce::VulkanTexture> vulkanTex =
                 ce::VulkanTexture::Create(ctx, "./assets/textures/" + textureNames[i]);
 
-            matToTex[i] = textureMng->allocTexture(vulkanTex);
+            matToTex[i] = static_cast<int>(textureMng->allocTexture(vulkanTex));
         }
     }
